Fix and improve message examples
The previous example for a `request` was incorrect. It had `"type": "quickinfo"`, which is not valid. In the process of fixing it, I also expanded it a little, to make a slightly more complete example.

Correspondingly, I changed the `"resposne"` example to show a response for a valid request as opposed to a response for an invalid request.